requirements.txt
asgiref==3.6.0
blinker==1.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
Deprecated==1.2.13
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==3.2.18
django-cors-headers==3.14.0
django-filter==23.1
django-mssql-backend==2.8.1
django-oauth-toolkit==2.2.0
django-pyodbc-azure==2.1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.6.0
ghp-import==2.1.0
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.2
jwcrypto==1.4.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.3.7
MarkupSafe==2.1.2
mergedeep==1.3.4
mkdocs==1.4.2
more-itertools==8.10.0
mssql-django==1.0rc1
netifaces==0.11.0
numpy==1.24.2
oauthlib==3.2.0
packaging==23.0
pandas==2.0.0
platformdirs==2.5.1
PyGObject==3.42.1
PyJWT==2.3.0
pyodbc==4.0.35
pyparsing==2.4.7
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==5.4.1
pyyaml_env_tag==0.1
requests==2.28.2
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.3
systemd-python==234
tzdata==2023.3
ubuntu-advantage-tools==27.11.2
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.15
virtualenv==20.13.0+ds
wadllib==1.3.6
watchdog==3.0.0
wrapt==1.15.0
zipp==1.0.0

# Instalar MSSQL ODBC UBUNTU
# https://learn.microsoft.com/pt-br/sql/connect/odbc/linux-mac/installing-the-microsoft-odbc-driver-for-sql-server?view=sql-server-ver16&tabs=ubuntu18-install%2Cubuntu17-install%2Cdebian8-install%2Credhat7-13-install%2Crhel7-offline
# if ! [[ "16.04 18.04 20.04 22.04" == *"$(lsb_release -rs)"* ]];
# then
#     echo "Ubuntu $(lsb_release -rs) is not currently supported.";
#     exit;
# fi

# sudo su
# curl https://packages.microsoft.com/keys/microsoft.asc | apt-key add -

# curl https://packages.microsoft.com/config/ubuntu/$(lsb_release -rs)/prod.list > /etc/apt/sources.list.d/mssql-release.list

# exit
# sudo apt-get update
# sudo ACCEPT_EULA=Y apt-get install -y msodbcsql17
# # optional: for bcp and sqlcmd
# sudo ACCEPT_EULA=Y apt-get install -y mssql-tools
# echo 'export PATH="$PATH:/opt/mssql-tools/bin"' >> ~/.bashrc
# source ~/.bashrc
# # optional: for unixODBC development headers
# sudo apt-get install -y unixodbc-dev
